Refactor : constants/browser 수정
상세 :
- client/config에 있던 LOCALSTORAGE_NICKNAME_KEY
  client/constant/browser로 이동

- client/constant/browser에 있던 COPY_TO_CLIPBOARD_MESSAGE
  client/constant/message로 이동

#388
