+++ client/src/constants/browser.js
@@ -1,11 +1,11 @@
 const INPUT_TAG = 'input';
 const BROWSER_COPY_COMMAND = 'copy';
-const COPY_TO_CLIPBOARD_MESSAGE = 'is copied to your clipboard';
 const ENTER_KEYCODE = 13;
+const LOCALSTORAGE_NICKNAME_KEY = 'trycatch-nickname';
 
 export {
   INPUT_TAG,
   BROWSER_COPY_COMMAND,
-  COPY_TO_CLIPBOARD_MESSAGE,
   ENTER_KEYCODE,
+  LOCALSTORAGE_NICKNAME_KEY,
 };
