+++ client/src/utils/shareUrlButtonClickHandler.js
@@ -1,6 +1,6 @@
 /* eslint-disable no-alert */
 import copyUrlToClipoard from './copyUrlToClipboard';
-import { COPY_TO_CLIPBOARD_MESSAGE } from '../constants/browser';
+import { COPY_TO_CLIPBOARD_MESSAGE } from '../constants/message';
 
 const shareUrlButtonClickHandler = () => {
   copyUrlToClipoard();
